backend/requirements/base.txt
Django==5.1.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
psycopg2-binary==2.9.10
channels==4.1.0
channels-redis==4.2.1
celery==5.4.0
redis==5.2.1
python-dotenv==1.0.1
gunicorn==23.0.0
daphne==4.1.2
structlog==25.1.0
whitenoise==6.8.2
